[Fix] sounddev: Fix 64bit compile.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@4769 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ sounddev/SoundDeviceWaveout.cpp
@@ -156,7 +156,7 @@ bool CWaveDevice::InternalOpen()
 	}
 	m_nWaveBufferSize = Util::Round<int32>(m_Settings.UpdateInterval * pwfx->nAvgBytesPerSec);
 	m_nWaveBufferSize = Util::AlignUp(m_nWaveBufferSize, pwfx->nBlockAlign);
-	m_nWaveBufferSize = Clamp(m_nWaveBufferSize, WAVEOUT_MINBUFFERFRAMECOUNT * pwfx->nBlockAlign, Util::AlignDown(WAVEOUT_MAXBUFFERSIZE, pwfx->nBlockAlign));
+	m_nWaveBufferSize = Clamp(m_nWaveBufferSize, static_cast<uint32>(WAVEOUT_MINBUFFERFRAMECOUNT * pwfx->nBlockAlign), static_cast<uint32>(Util::AlignDown(WAVEOUT_MAXBUFFERSIZE, pwfx->nBlockAlign)));
 	std::size_t numBuffers = Util::Round<int32>(m_Settings.Latency * pwfx->nAvgBytesPerSec / m_nWaveBufferSize);
 	numBuffers = Clamp(numBuffers, WAVEOUT_MINBUFFERS, WAVEOUT_MAXBUFFERS);
 	m_nPreparedHeaders = 0;
